+++ instrumentation/opentelemetry-instrumentation-urllib3/tests/test_urllib3_metrics.py
@@ -14,13 +14,30 @@
 
 import io
 from timeit import default_timer
+from unittest import mock
 
 import httpretty
 import urllib3
 import urllib3.exceptions
 from urllib3 import encode_multipart_formdata
 
+from opentelemetry.instrumentation._semconv import (
+    OTEL_SEMCONV_STABILITY_OPT_IN,
+    _OpenTelemetrySemanticConventionStability,
+)
 from opentelemetry.instrumentation.urllib3 import URLLib3Instrumentor
+from opentelemetry.semconv.attributes.http_attributes import (
+    HTTP_REQUEST_METHOD,
+    HTTP_RESPONSE_STATUS_CODE,
+)
+from opentelemetry.semconv.attributes.network_attributes import (
+    NETWORK_PROTOCOL_VERSION,
+)
+from opentelemetry.semconv.attributes.server_attributes import (
+    SERVER_ADDRESS,
+    SERVER_PORT,
+)
+from opentelemetry.semconv.trace import SpanAttributes
 from opentelemetry.test.httptest import HttpTestBase
 from opentelemetry.test.test_base import TestBase
 
@@ -30,6 +47,24 @@ class TestURLLib3InstrumentorMetric(HttpTestBase, TestBase):
 
     def setUp(self):
         super().setUp()
+
+        test_name = ""
+        if hasattr(self, "_testMethodName"):
+            test_name = self._testMethodName
+        sem_conv_mode = "default"
+        if "new_semconv" in test_name:
+            sem_conv_mode = "http"
+        elif "both_semconv" in test_name:
+            sem_conv_mode = "http/dup"
+
+        self.env_patch = mock.patch.dict(
+            "os.environ",
+            {
+                OTEL_SEMCONV_STABILITY_OPT_IN: sem_conv_mode,
+            },
+        )
+        _OpenTelemetrySemanticConventionStability._initialized = False
+        self.env_patch.start()
         URLLib3Instrumentor().instrument()
         httpretty.enable(allow_net_connect=False)
         httpretty.register_uri(httpretty.GET, self.HTTP_URL, body="Hello!")
@@ -38,6 +73,7 @@ class TestURLLib3InstrumentorMetric(HttpTestBase, TestBase):
 
     def tearDown(self):
         super().tearDown()
+        self.env_patch.stop()
         self.pool.clear()
         URLLib3Instrumentor().uninstrument()
 
@@ -47,8 +83,7 @@ class TestURLLib3InstrumentorMetric(HttpTestBase, TestBase):
     def test_basic_metrics(self):
         start_time = default_timer()
         response = self.pool.request("GET", self.HTTP_URL)
-        client_duration_estimated = (default_timer() - start_time) * 1000
-
+        duration_ms = max(round((default_timer() - start_time) * 1000), 0)
         metrics = self.get_sorted_metrics()
 
         (
@@ -57,27 +92,29 @@ class TestURLLib3InstrumentorMetric(HttpTestBase, TestBase):
             client_response_size,
         ) = metrics
 
+        attrs_old = {
+            SpanAttributes.HTTP_STATUS_CODE: 200,
+            SpanAttributes.HTTP_HOST: "mock",
+            SpanAttributes.NET_PEER_PORT: 80,
+            SpanAttributes.NET_PEER_NAME: "mock",
+            SpanAttributes.HTTP_METHOD: "GET",
+            SpanAttributes.HTTP_FLAVOR: "1.1",
+            SpanAttributes.HTTP_SCHEME: "http",
+        }
+
         self.assertEqual(client_duration.name, "http.client.duration")
         self.assert_metric_expected(
             client_duration,
             [
                 self.create_histogram_data_point(
                     count=1,
-                    sum_data_point=client_duration_estimated,
-                    max_data_point=client_duration_estimated,
-                    min_data_point=client_duration_estimated,
-                    attributes={
-                        "http.flavor": "1.1",
-                        "http.host": "mock",
-                        "http.method": "GET",
-                        "http.scheme": "http",
-                        "http.status_code": 200,
-                        "net.peer.name": "mock",
-                        "net.peer.port": 80,
-                    },
+                    sum_data_point=duration_ms,
+                    max_data_point=duration_ms,
+                    min_data_point=duration_ms,
+                    attributes=attrs_old,
                 )
             ],
-            est_value_delta=200,
+            est_value_delta=40,
         )
 
         self.assertEqual(client_request_size.name, "http.client.request.size")
@@ -89,15 +126,7 @@ class TestURLLib3InstrumentorMetric(HttpTestBase, TestBase):
                     sum_data_point=0,
                     max_data_point=0,
                     min_data_point=0,
-                    attributes={
-                        "http.flavor": "1.1",
-                        "http.host": "mock",
-                        "http.method": "GET",
-                        "http.scheme": "http",
-                        "http.status_code": 200,
-                        "net.peer.name": "mock",
-                        "net.peer.port": 80,
-                    },
+                    attributes=attrs_old,
                 )
             ],
         )
@@ -114,15 +143,225 @@ class TestURLLib3InstrumentorMetric(HttpTestBase, TestBase):
                     sum_data_point=expected_size,
                     max_data_point=expected_size,
                     min_data_point=expected_size,
-                    attributes={
-                        "http.flavor": "1.1",
-                        "http.host": "mock",
-                        "http.method": "GET",
-                        "http.scheme": "http",
-                        "http.status_code": 200,
-                        "net.peer.name": "mock",
-                        "net.peer.port": 80,
-                    },
+                    attributes=attrs_old,
+                )
+            ],
+        )
+
+    def test_basic_metrics_new_semconv(self):
+        start_time = default_timer()
+        response = self.pool.request("GET", self.HTTP_URL)
+        duration_s = max(default_timer() - start_time, 0)
+
+        metrics = self.get_sorted_metrics()
+        (
+            client_request_size,
+            client_duration,
+            client_response_size,
+        ) = metrics
+
+        attrs_new = {
+            NETWORK_PROTOCOL_VERSION: "1.1",
+            SERVER_ADDRESS: "mock",
+            SERVER_PORT: 80,
+            HTTP_REQUEST_METHOD: "GET",
+            HTTP_RESPONSE_STATUS_CODE: 200,
+            # TODO: add URL_SCHEME to tests when supported in the implementation
+        }
+
+        self.assertEqual(client_duration.name, "http.client.request.duration")
+        self.assert_metric_expected(
+            client_duration,
+            [
+                self.create_histogram_data_point(
+                    count=1,
+                    sum_data_point=duration_s,
+                    max_data_point=duration_s,
+                    min_data_point=duration_s,
+                    attributes=attrs_new,
+                )
+            ],
+            est_value_delta=40 / 1000,
+        )
+
+        self.assertEqual(
+            client_request_size.name, "http.client.request.body.size"
+        )
+        self.assert_metric_expected(
+            client_request_size,
+            [
+                self.create_histogram_data_point(
+                    count=1,
+                    sum_data_point=0,
+                    max_data_point=0,
+                    min_data_point=0,
+                    attributes=attrs_new,
+                )
+            ],
+        )
+
+        expected_size = len(response.data)
+        self.assertEqual(
+            client_response_size.name, "http.client.response.body.size"
+        )
+        self.assert_metric_expected(
+            client_response_size,
+            [
+                self.create_histogram_data_point(
+                    count=1,
+                    sum_data_point=expected_size,
+                    max_data_point=expected_size,
+                    min_data_point=expected_size,
+                    attributes=attrs_new,
+                )
+            ],
+        )
+
+    @mock.patch("httpretty.http.HttpBaseClass.METHODS", ("NONSTANDARD",))
+    def test_basic_metrics_nonstandard_http_method(self):
+        httpretty.register_uri(
+            "NONSTANDARD", self.HTTP_URL, body="", status=405
+        )
+
+        start_time = default_timer()
+        response = self.pool.request("NONSTANDARD", self.HTTP_URL)
+        duration_ms = max(round((default_timer() - start_time) * 1000), 0)
+
+        metrics = self.get_sorted_metrics()
+
+        (
+            client_duration,
+            client_request_size,
+            client_response_size,
+        ) = metrics
+
+        attrs_old = {
+            SpanAttributes.HTTP_STATUS_CODE: 405,
+            SpanAttributes.HTTP_HOST: "mock",
+            SpanAttributes.NET_PEER_PORT: 80,
+            SpanAttributes.NET_PEER_NAME: "mock",
+            SpanAttributes.HTTP_METHOD: "_OTHER",
+            SpanAttributes.HTTP_FLAVOR: "1.1",
+            SpanAttributes.HTTP_SCHEME: "http",
+        }
+
+        self.assertEqual(client_duration.name, "http.client.duration")
+        self.assert_metric_expected(
+            client_duration,
+            [
+                self.create_histogram_data_point(
+                    count=1,
+                    sum_data_point=duration_ms,
+                    max_data_point=duration_ms,
+                    min_data_point=duration_ms,
+                    attributes=attrs_old,
+                )
+            ],
+            est_value_delta=40,
+        )
+
+        self.assertEqual(client_request_size.name, "http.client.request.size")
+        self.assert_metric_expected(
+            client_request_size,
+            [
+                self.create_histogram_data_point(
+                    count=1,
+                    sum_data_point=0,
+                    max_data_point=0,
+                    min_data_point=0,
+                    attributes=attrs_old,
+                )
+            ],
+        )
+
+        expected_size = len(response.data)
+        self.assertEqual(
+            client_response_size.name, "http.client.response.size"
+        )
+        self.assert_metric_expected(
+            client_response_size,
+            [
+                self.create_histogram_data_point(
+                    count=1,
+                    sum_data_point=expected_size,
+                    max_data_point=expected_size,
+                    min_data_point=expected_size,
+                    attributes=attrs_old,
+                )
+            ],
+        )
+
+    @mock.patch("httpretty.http.HttpBaseClass.METHODS", ("NONSTANDARD",))
+    def test_basic_metrics_nonstandard_http_method_new_semconv(self):
+        httpretty.register_uri(
+            "NONSTANDARD", self.HTTP_URL, body="", status=405
+        )
+        start_time = default_timer()
+        response = self.pool.request("NONSTANDARD", self.HTTP_URL)
+        duration_s = max(default_timer() - start_time, 0)
+
+        metrics = self.get_sorted_metrics()
+
+        (
+            client_request_size,
+            client_duration,
+            client_response_size,
+        ) = metrics
+
+        attrs_new = {
+            NETWORK_PROTOCOL_VERSION: "1.1",
+            SERVER_ADDRESS: "mock",
+            SERVER_PORT: 80,
+            HTTP_REQUEST_METHOD: "_OTHER",
+            HTTP_RESPONSE_STATUS_CODE: 405,
+            "error.type": "405",
+            # TODO: add URL_SCHEME to tests when supported in the implementation
+        }
+
+        self.assertEqual(client_duration.name, "http.client.request.duration")
+        self.assert_metric_expected(
+            client_duration,
+            [
+                self.create_histogram_data_point(
+                    count=1,
+                    sum_data_point=duration_s,
+                    max_data_point=duration_s,
+                    min_data_point=duration_s,
+                    attributes=attrs_new,
+                )
+            ],
+            est_value_delta=40 / 1000,
+        )
+
+        self.assertEqual(
+            client_request_size.name, "http.client.request.body.size"
+        )
+        self.assert_metric_expected(
+            client_request_size,
+            [
+                self.create_histogram_data_point(
+                    count=1,
+                    sum_data_point=0,
+                    max_data_point=0,
+                    min_data_point=0,
+                    attributes=attrs_new,
+                )
+            ],
+        )
+
+        expected_size = len(response.data)
+        self.assertEqual(
+            client_response_size.name, "http.client.response.body.size"
+        )
+        self.assert_metric_expected(
+            client_response_size,
+            [
+                self.create_histogram_data_point(
+                    count=1,
+                    sum_data_point=expected_size,
+                    max_data_point=expected_size,
+                    min_data_point=expected_size,
+                    attributes=attrs_new,
                 )
             ],
         )
@@ -274,3 +513,5 @@ class TestURLLib3InstrumentorMetric(HttpTestBase, TestBase):
         for metric in metrics:
             for point in list(metric.data.data_points):
                 self.assertEqual(point.count, 1)
+        # instrument again to avoid warning message on tearDown
+        URLLib3Instrumentor().instrument()